controllers and endpoints created

+++ database/migrations/2026_05_08_202429_create_transactions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained();
             $table->foreignId('category_id')->constrained();
             $table->string('description')->nullable();
-            $table->decimal('amount', 10, 2);
+            $table->unsignedBigInteger('amount');
             $table->enum('type', ['income', 'expense'])->default('expense');
             $table->timestamps();
         });
